// src/api/rest/ingredients/type.ts
export interface IIngredient {
  _id: string;
  name: string;
  type: string;
  proteins: number;
  fat: number;
  carbohydrates: number;
  calories: number;
  price: number;
  image: string;
  image_mobile: string;
  image_large: string;
  __v: number;
  dragId: string;
}

export interface IIngredients {
  success: string;
  data: IIngredient[];
}